work on integration tests, general quality fixes

+++ stature.py
@@ -55,33 +55,36 @@ def main(cli, cach, settings):
                 logging.info("Creating Component: %s", args['name'])
                 # assume status is fine
                 ret = cach.postComponents(status=1, **args)
+                # print(ret.json())
                 ret.raise_for_status()
                 cach_id = ret.json()['data']['id']
-                logging.info("Got component id: %d", cach_id)
+                logging.info("Got component id: %s", cach_id)
                 settings['containers'][name] = cach_id
         else:
             logging.info(
                 "Container: %s not found in your toml file, nor does it have a docker label metadata, see the docs for refrence.", name)
             continue
         status = container.status
-        # time.sleep(2)
-        # Why was this here. slowing tests???
-        logging.debug("Cachet ID: %d",cach_id)
-        if status == "running":
-            ret = cach.putComponentsByID(cach_id, status=1)
-            logging.debug(ret.text)
-            ret.raise_for_status()
-        elif status == "exited":
-            ret = cach.putComponentsByID(cach_id, status=4)
-            ret.raise_for_status()
+        logging.debug("Cachet ID: %s",cach_id)
+        logging.debug("container: %s", container)
+        if cach_id:
+            if status == "running":
+                ret = cach.putComponentsByID(cach_id, status=1)
+                logging.debug(ret.text)
+                ret.raise_for_status()
+            elif status == "exited":
+                ret = cach.putComponentsByID(cach_id, status=4)
+                ret.raise_for_status()
     return settings
 
 
 @click.command(context_settings=CONTEXT_SETTINGS)
+# @click.option("--all", '-a', default=False, is_flag=True, help="Check exited containers too")
+@click.option("--keep", "-k", default=False, is_flag=True, help="Keep cachet ids in your toml file")
 @click.option('--debug', default=False, is_flag=True)
 @click.option('--verbose/--silent', default=True)
 @click.option("--conf-file", '-f', default="docker2cachet.toml", type=click.Path(exists=True))
-def run(conf_file, verbose, debug):
+def run(conf_file, verbose, debug, keep):
     settings = toml.load(conf_file)
     if debug:
         logging.basicConfig(level=logging.DEBUG)
@@ -90,8 +93,9 @@ def run(conf_file, verbose, debug):
     cli = docker.from_env()
     cach = Cachet(settings['cachet']['url'], settings['cachet']['api_key'])
     settings = main(cli, cach, settings)
-    with open(conf_file, 'w') as f:
-        toml.dump(settings, f)
+    if keep:
+        with open(conf_file, 'w') as f:
+            toml.dump(settings, f)
 
 if __name__ == '__main__':
     run()
